requirements.txt
btrees==4.2.0
chameleon==2.24
cornice==1.2.0
csscompressor==0.9.4
docutils==0.12
funcsigs==0.4
HTMLParser==0.0.2
jsbeautifier==1.6.2
mako==1.0.3
markupsafe==0.23
mock==1.3.0
pastedeploy==1.5.2
pbr==1.8.1
persistent==4.1.1
ply==3.8
pygments==2.1.3
pyramid==1.6.1
pyramid-chameleon==0.3
pyramid-debugtoolbar==2.4.2
pyramid-mako==1.0.2
pyramid-tm==0.12.1
pyramid-zodbconn==0.7
PyYAML==3.11
repoze.lru==0.6
simplejson==3.8.2
six==1.10.0
slimit==0.8.1
transaction==1.4.4
translationstring==1.3
venusian==1.0
waitress==0.8.10
webob==1.6.0a0
wheel==0.24.0
zc.lockfile==1.1.0
zconfig==3.1.0
zdaemon==4.1.0
zeo==4.2.0b1
zodb==4.2.0
zodb3==3.11.0
zodbpickle==0.6.0
zodburi==2.0
zope.deprecation==4.1.2
zope.interface==4.1.3
